microservice/requirements.txt
kafka-python
gensim
nltk
